pyproject.toml

[project]
name = "openapi-tools"
authors = [
  { name = "openapi-tools", email = "hello@fvoron.com" }
]
description = "Python tools to work with OpenAPI schemas"
readme = "README.md"
dynamic = ["version"]
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "openapi-pydantic>=0.5.1",
    "pyyaml>=6.0.3",
    "textual>=8.0.0",
]

[project.urls]
Documentation = "https://frankie567.github.io/openapi-tools/"
Source = "https://github.com/frankie567/openapi-tools"

[build-system]
requires = ["hatchling", "hatch-regex-commit"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "regex_commit"
commit_extra_args = ["-e"]
path = "openapi_tools/__init__.py"

[dependency-groups]
dev = [
    "mypy",
    "ruff",
    "pytest",
    "pytest-cov",
    "mkdocs-material",
    "mkdocstrings[python]",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
extend-select = ["I", "TRY", "UP"]

[tool.pytest.ini_options]
addopts = "--cov=openapi_tools/ --cov-report=term-missing"


[tool.uv]
default-groups = ["dev"]

[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
strict_equality = true
disallow_untyped_defs = true

[tool.basedpyright]
typeCheckingMode = "off"
